MemberAPI, AlarmAPI, AuthAPI 추가
